drugbox page to map view
