+++ waste_drug/app/src/main/AndroidManifest.xml
@@ -14,16 +14,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Waste_drug"
         android:usesCleartextTraffic="true">
-        <activity android:name=".KakaoActivity"></activity>
+        <activity android:name=".Map1Activity"></activity>
+        <activity android:name=".Map2Activity" />
         <activity android:name=".search.SearchActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
         </activity>
-
 
         <meta-data
             android:name="com.kakao.sdk.AppKey"
