+++ waste_drug/app/src/main/java/com/example/waste_drug/db/DrugBox.java
@@ -3,9 +3,10 @@ package com.example.waste_drug.db;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
+import java.io.Serializable;
 
 @Entity
-public class DrugBox {
+public class DrugBox implements Serializable{
     @PrimaryKey(autoGenerate = true)
     public int uid = 0;
